fix: create common EventModel for event
