+++ src/components/search-result/EventList/EventCard/EventCard.tsx
@@ -2,9 +2,9 @@ import * as s from './EventCard.styles';
 import { useTranslation } from 'react-i18next';
 import { useState } from 'react';
 import { StarOutlined, StarFilled } from '@ant-design/icons';
-import { FilterOutlined, CalendarOutlined } from "@ant-design/icons";
+import EventModel from '@/domain/EventModel';
 
-const EventCard: React.FC<s.EventProps> = ({
+const EventCard: React.FC<EventModel> = ({
   id,
   eventName,
   eventLogoURL,
